build(docs): move from docz to parcel for docs
