--- demo/App.mdx
@@ -1,8 +1,0 @@
----
-name: Demo
-route: /
----
-
-import App from "./App";
-
-<App />
