+++ demo/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="utf-8" />
+        <title>My First Parcel App</title>
+        <script type="module" src="index.jsx"></script>
+    </head>
+    <body>
+        <div id="root" />
+    </body>
+</html>
