+++ demo/index.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+import ReactDOM from "react-dom";
+import { App } from "./App";
+
+ReactDOM.render(<App />, document.getElementById("root"));
