--- doczrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-    base: "/react-geolocated/",
-    ignore: [
-        "CHANGELOG.md",
-        "CODE_OF_CONDUCT.md",
-        "CONTRIBUTING.md",
-        "LICENSE",
-    ],
-};
